requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
stripe
python-dotenv
